+++ Runtime/SDK/RelayService.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Runtime.CompilerServices;
+using Unity.Services.Authentication.Internal;
 #if USE_QOS
 using Unity.Services.Relay.Qos;
 #endif
@@ -11,16 +13,17 @@ namespace Unity.Services.Relay
     /// <summary>
     /// The entry class of the Relay Allocations Service enables clients to connect to relay servers. Once connected, they are able to communicate with each other, via the relay servers, using the bespoke relay binary protocol.
     /// </summary>
-    public static class Relay
+    public static class RelayService
     {
-        private static IRelayServiceSDK service;
+        private static IRelayService service;
         
         private static readonly Configuration allocationsApiConfiguration;
+
 #if USE_QOS
         private static readonly Qos.Configuration qosDiscoveryApiConfiguration;
 #endif
 
-        static Relay()
+        static RelayService()
         {
 #if AUTHENTICATION_TESTING_STAGING_UAS
             allocationsApiConfiguration = new Configuration("https://relay-allocations-stg.services.api.unity.com", 10, 4, null);
@@ -38,22 +41,32 @@ namespace Unity.Services.Relay
         /// <summary>
         /// A static instance of the Relay Allocation Client.
         /// </summary>
-        public static IRelayServiceSDK Instance
+        public static IRelayService Instance
         {
             get
             {
-                if (service != null) return service;
+                if (service != null)
+                {
+                    return service;
+                }
 
-#if USE_QOS
-                IQosService qosService = new RelayQosService(QosDiscoveryService.Instance.QosDiscoveryApi, qosDiscoveryApiConfiguration, new MultiplayAdapterQosRunner());
-#else
-                IQosService qosService = null;
-#endif
-                service = new WrappedRelayService(RelayService.Instance.AllocationsApi, allocationsApiConfiguration,
-                    qosService);
+                service = new WrappedRelayService(RelayServiceSdk.Instance);
 
                 return service;
             }
         }
+    }
+
+    /// <summary>
+    /// "Relay class is marked for deprecation. Please use RelayService class instead."
+    /// </summary>
+    //[Obsolete("Relay class is marked for deprecation. Please use RelayService class instead.")]
+    public static class Relay
+    {
+        /// <summary>
+        /// Relay class is marked for deprecation. Please use RelayService.Instance instead.
+        /// </summary>
+        //[Obsolete("Relay class is marked for deprecation. Please use RelayService.Instance instead.")]
+        public static IRelayServiceSDK Instance { get { return (IRelayServiceSDK) RelayService.Instance; } }
     }
 }